[Build] restore osdVersion to 2.0.0
By keeping `opensearchDashboardsVersion: 2.0.0`, plugin owners can
do development using the main branch of OpenSearch Dashboards since
OpenSearch Dashboards sets this value to 2.0.0.

This key-value lets OpenSearch Dashboards can properly install the
plugin since it does a semvar + qualifier check on this value if
they match then OpenSearch Dashboards proceeds to install it.

The build repo passes the qualifier while building OpenSearch Dashboards
and plugins that automatically sets this value to build the correct
version + qualifier. This allows a distribution to be made with value
for `opensearchDashboardsVersion` be the expected value, for example
`opensearchDashboardsVersion: 2.0.0-rc1`.

Signed-off-by: Kawika Avilla <kavilla414@gmail.com>