Select tenant popup only appears when mutli-tenacy is enabled (#965)
The current behavoir is that all users will see a message to switch
tenants and if multi-tenancy is disabled, that message will be prompting
the user to ask the admin to enable the feature.  This is causing users
pain that are not interested in using the tenancy features.

Signed-off-by: Peter Nied <petern@amazon.com>
(cherry picked from commit ec59de0edfc30ecc12219c3259f8869fc609d803)