Remove multi-tenant path check in auth handler.
The Opensearch Dashboards tenant only impacts the saved objects access. The multi-tenant path check in the auth handler
was meant to by pass the resolve tenant logic for some APIs which doesn't deal with saved objects. However with the new
datasource support feature in OpenSearch Dashboards, Dashboards APIs which used to not dealing with saved objects such
as data queries and index resolve operations will require the new data source saved object type under the hood.

This change removes the multi-tenant path check, so that tenant header will be injected to all route handlers. So that
all route handlers can work with the right tenant in case they need to deal with any saved objects.

Signed-off-by: Yan Zeng <zengyan@amazon.com>