Add retry and enable UTs for Windows
Added a retry around the the bootstrap that previously failed due to network issues.  Ran the new version of this workflow 100x on linux and 100x on windows, both passed at 100% [1].

Without the retry during a run of 100 times iterations [2], the windows test failed once during the bootstrap.  If we see future issues, we could attempt to add retries to them following this same pattern.

[1] https://github.com/peternied/security-dashboards-plugin/actions/runs/3449518073
[2] https://github.com/peternied/security-dashboards-plugin/actions/runs/3449040181

- Resolves #1211
- Resolves #1195

Signed-off-by: Peter Nied <petern@amazon.com>